+++ src/test/java/shopping/config/TestConfig.java
@@ -6,8 +6,6 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.http.client.BufferingClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.client.RestTemplate;
 import shopping.infrastructure.CurrencyLayerExchangeRateApi;
 import shopping.infrastructure.ExchangeRateApi;
@@ -37,8 +35,6 @@ public class TestConfig {
     public RestTemplate restTemplate(final RestTemplateBuilder restTemplateBuilder) {
 
         return restTemplateBuilder
-            .requestFactory(
-                () -> new BufferingClientHttpRequestFactory(new SimpleClientHttpRequestFactory()))
             .setConnectTimeout(Duration.ofMillis(CONNECTION_TIME_OUT))
             .setReadTimeout(Duration.ofMillis(READ_TIME_OUT))
             .build();
